% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_strategy_attack.R
\name{adaptive_strategy_attack}
\alias{adaptive_strategy_attack}
\title{Make Strategy Adaptive}
\usage{
adaptive_strategy_attack(g, attack_strategy, target = "nodes",
  number = 1)
}
\arguments{
\item{g}{The network which is being attacked}

\item{number}{The number of targets to generate}

\item{nodes}{The number of nodes to Remove}

\item{seed}{The random number generator seed to use. This is optional}
}
\description{
This implements an attack strategy to be adaptive.
    This means that it will be re-evaluated every
    round to choose the most effective point of attack
    according to that metric
}
\examples{
adaptive_strategy_attack <- function(g, attack_strategy, target = "nodes", number = 1)
}
\keyword{regime}
\keyword{removal}
