% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_the_grid.R
\name{attack_the_grid}
\alias{attack_the_grid}
\title{Initiate power-grid attack simulation}
\usage{
attack_the_grid(network_list, attack_strategy, reference_grid = NULL,
  min_max_comp = 0, total_attack_rounds = 1000, cascade_mode = TRUE,
  cumulative_attacks = NULL, demand = "demand",
  generation = "generation", edge_name = "edge_name",
  node_name = "name", net_generation = "net_generation")
}
\arguments{
\item{network_list}{A list of lists where each element of the sub list is an igraph object, the first time it is used the
the network list is simply list(list(g)).}

\item{attack_strategy}{A function that calculates which node to delete the function is is in "quo" form and embedded in an
attack type.}

\item{reference_grid}{The grid that will be used to test the largest component against if NULL it uses the given network.}

\item{min_max_comp}{The minimum size of the maximum component, as a fraction, for the process to continue, the default is set
to 0.0 complete collapse.}

\item{total_attack_rounds}{The maximum number of nodes to be removed before the process stops.}

\item{cascade_mode}{Whether the power flow equations will be used to check line-overloading or not.}

\item{cumulative_attacks}{The total number of attacks that have taken place so far.}

\item{demand}{the node_name of the node Load variable. A character string.}

\item{generation}{the node_name of the node generation variable. A character string.}

\item{edge_name}{the variable that holds the edge names, a character string.}

\item{node_name}{the variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{net_generation}{the node_name that the net generation data for each node is held in}
}
\description{
This function simulates an attack on the power grid using the parameter settings you choose
   the outut of the function is a nested list of igraph objects.
}
\examples{
attack_the_grid(network_list, attack_strategy, SubstationData, EdgeData, reference_grid = NULL, min_max_comp = 0.8)
Out <- attack_the_grid(network_list,
attack_strategy,
reference_grid = NULL,
min_max_comp = 0.8,
total_attack_rounds=100,
cascade_mode = TRUE,
cumulative_attacks = NULL)
}
