% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttackTheGrid.R
\name{AttackTheGrid}
\alias{AttackTheGrid}
\title{Initiate power-grid attack simulation}
\usage{
AttackTheGrid(NetworkList, AttackStrategy, SubstationData, EdgeData,
  referenceGrid = NULL, MinMaxComp = 0.8)
}
\arguments{
\item{NetworkList}{a list of lists of networks, when initiating the attack use list(list(g))}

\item{AttackStrategy}{A function that calculates which node to delete the function is is in "quo" form}

\item{SubstationData}{Dataframe that contains data on the susbtations}

\item{EdgeData}{Data frame of Edge data for the network.}

\item{referenceGrid}{the grid that will be used to test the largest component against if NULL it uses the given network}

\item{MinMaxComp}{The minimum size of the maximum component for the process to continue}
}
\description{
This function simulates an attack on the power grid using the parameter settings you choose
   the outut of the function is a nested list of igraph objects.
}
