% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_generation_and_demand.R
\name{balance_generation_and_demand}
\alias{balance_generation_and_demand}
\title{Balance Power network}
\usage{
balance_generation_and_demand(g, demand, generation,
  output_var = "net_generation")
}
\arguments{
\item{g}{An igraph object representing a power network. A character vector.}

\item{demand}{The variable in g that contains the demand of each node. A character vector.}

\item{generation}{The variable in g that contains the generation of each node. A character vector.}

\item{output_var}{The desired output variable for the balanced net power demand of each node.
A character vector the default is set to "net_generation".}
}
\description{
Ensures that the demand and generation of the power network are balanced. This is used before power
  flow is calculated.
}
\examples{
balance_generation_and_demand(g, demand, generation, output_var = "net_generation")
}
