% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalancedGenDem.R
\name{BalancedGenDem}
\alias{BalancedGenDem}
\title{Balance Power network}
\usage{
BalancedGenDem(g, DemandVar, GenerationVar, OutputVar = "BalencedPower")
}
\arguments{
\item{g}{An igraph object representing a power network. A character vector.}

\item{DemandVar}{The variable in g that contains the demand of each node. A character vector.}

\item{GenerationVar}{The variable in g that contains the generation of each node. A character vector.}

\item{OutputVar}{The desired output variable for the balanced net power demand of each node.
A character vector the default is set to "BalencedPower".}
}
\description{
Ensures that the demand and generation of the power network are balanced. This is used before power
  flow is calculated.
}
\examples{
BalencedGenDem(g, DemandVar, GenerationVar, OutputVar = "BalencedPower")
}
