% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biggest_collapse.R
\name{biggest_collapse}
\alias{biggest_collapse}
\title{Biggest Collapse}
\usage{
biggest_collapse(df, var)
}
\arguments{
\item{df}{A data frame. A dataframe describing the change in grid structure across attacks. Either the output of extract_attack_stats
or extract_network_stats}

\item{var}{bare unquoted variable node_name. Where you are interested in finding the biggest drop}
}
\description{
This function finds largest drop in performance for a given metric. Specifically it looks for the lowest
number in the given variable across all simulations and arranges them in rank order.
}
\examples{
biggest_collapse(df, alpha)
}
\seealso{
\code{\link{extract_network_stats}}, \code{\link{extract_attack_stats}}
}
