% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_over_limit.R
\name{calc_over_limit}
\alias{calc_over_limit}
\title{Find which edges are over the limit}
\usage{
calc_over_limit(g, edge_name = "edge_name", node_name = "name",
  net_generation = "net_generation")
}
\arguments{
\item{g}{An igraph object that represents a power-grid}

\item{edge_name}{the variable that holds the edge names, a character string.}

\item{node_name}{the variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{net_generation}{the node_name that the net generation data for each node is held in}
}
\description{
This function checks to see which lines are over limit and removes them and re-calculates power flow in the network
}
\details{
The function uses a node attribute "bus_order" to identify the slack bus. This may be changed in later versions
}
