% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_and_elecdist.R
\name{capacity_and_elecdist}
\alias{capacity_and_elecdist}
\title{Capacity and electrical distance}
\usage{
capacity_and_elecdist(g, slack_ref)
}
\arguments{
\item{g}{An igraph object}

\item{slack_ref}{the reference bus}
}
\description{
This is a wrapper that makes the code for Net ability easier to read.
It only calculates the transfer capacity of the network and the electrical distance
}
