% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cascade.R
\name{Cascade}
\alias{Cascade}
\title{Calculate the effects of a cascade}
\usage{
Cascade(NetworkList, Iteration = 0, StopCascade = Inf, g0 = NULL)
}
\arguments{
\item{NetworkList}{A list of power-grid networks.}

\item{Iteration}{The current iteration number}

\item{StopCascade}{The number of iterations when cascade will be forced to terminate. An integer, default is set to infinity.}
}
\description{
Iterates through the network removing edges until no edges are over the line limit and the network has stabilised.
}
\details{
This is a recursive function that calls itself repeatedly until the cascade has come to a stop and the network has stabilised.
  The function produces a list of graphs showing each state of the cascade. This function is mostly used as part of the
  AttackTheGrid function.
}
\seealso{
\code{\link{AttackTheGrid}}
}
