% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cascade2.R
\name{Cascade2}
\alias{Cascade2}
\title{Calculate the effects of a cascade}
\usage{
Cascade2(
  g,
  g0 = NULL,
  node_power = node_power,
  edge_status = edge_status,
  AZero,
  LineProperties,
  Generation = "Generation",
  Demand = "Demand",
  EdgeName = "Link",
  VertexName = "name",
  Net_generation = "BalencedPower",
  power_flow = "PowerFlow",
  edge_limit = "Link.Limit"
)
}
\arguments{
\item{g}{An igraph object. The graph at its current state in the attack process}

\item{g0}{An igraph object. An optional value that defines the reference network}

\item{node_power}{A vector of the current node power for this round}

\item{edge_status}{A vector of current edge status, this is an integer vector of 4 levels 0-3}

\item{AZero}{A matrix. THe transmission matrix of the network}

\item{LineProperties}{The diagonal Line properties matrix}

\item{Generation}{The name of the node generation variable. A character string.}

\item{Demand}{the name of the node Load variable. A character string.}

\item{EdgeName}{The variable that holds the edge names, a character string.}

\item{VertexName}{The variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{Net_generation}{The name that the net generation data for each node is held in}

\item{power_flow}{A character string. This value indicates the name of the edge attribute that holds power flow, the default is "PowerFlow"}

\item{edge_limit}{A character string. This value indicates the name of the edge attribute that holds the edge limit, the default is "Link.Limit"}
}
\description{
Iterates through the network removing edges until no edges are over the line limit and the network has stabilised.
}
\details{
This version of cascade works with the attack_the_grid function. It is confusing as it is called cascade2 not one but there we go.

The function outputs a list of 3 elements. the first element is the updated igraph oobject after the cascade has terminated.
The second element is a vector of node power values, the third element is a vector of the edge status.
}
\seealso{
\code{\link{attack_the_grid}}
}
