% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Components_differ.R
\name{Components_differ}
\alias{Components_differ}
\title{Find components of a graph that have
Find components of a graph that have changed}
\usage{
Components_differ(g, g0)
}
\arguments{
\item{g}{the current graph}

\item{g0}{the previous graph}
}
\description{
Using this should be able to cut down on the amount of calculation done on the power flow equations
}
