% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subgraph.R
\name{create_subgraph}
\alias{create_subgraph}
\title{Create Subgraph}
\usage{
create_subgraph(g, nodes, edge_limit = "edge_limit")
}
\arguments{
\item{g}{and igraph object}

\item{nodes}{A character vector of node names to be in the new graph}

\item{edge_limit}{The graph attribute node_name for the line limits of the network}
}
\description{
This function creates a subgraph where the nodes which have lost links to outside the subgraph gain additional generation
capacity. This generation capacity is limited to the line limits of the PowerLines.
This limit prevents those lines from being overpowered, although it assumes such power is actually available
}
