% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transmission.R
\name{create_transmission}
\alias{create_transmission}
\title{Create Transmission Network}
\usage{
create_transmission(g, edge_name = "edge_name", node_name = "name")
}
\arguments{
\item{g}{an igraph object with attribute of named edges. This is an igraph object}

\item{edge_name}{The attribute node_name of the edges. This is a string}

\item{node_name}{the variable that holds the names of the nodes. a character string.}
}
\description{
This function creates the transmission matrix between the edges and the In and Out nodes
}
\details{
The function follows the method described in Pepyne 2007.
   It takes a graph an ouputs a matrix which has -1,0-1 values which describe the nominal direction of flow.
   The function is typically used in conjunction with the line_properties_matrix function. The function
   is a key part of calculating imp_ptdf.
}
\examples{
 create_transmission(g,  edge_name = "edge_name", node_name = "name")
}
\seealso{
\code{\link{line_properties_matrix}} \code{\link{imp_ptdf}}
}
