% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTransmission.R
\name{CreateTransmission}
\alias{CreateTransmission}
\title{Create Transmission Network}
\usage{
CreateTransmission(g, EdgeName = "Link", VertexName = "name")
}
\arguments{
\item{g}{an igraph object with attribute of named edges. This is an igraph object}

\item{EdgeName}{The attribute name of the edges. This is a string}

\item{VertexName}{the variable that holds the names of the nodes. a character string.}
}
\description{
This function creates the transmission matrix between the edges and the In and Out Nodes
}
\details{
The function follows the method described in Pepyne 2007.
   It takes a graph an ouputs a matrix which has -1,0-1 values which describe the nominal direction of flow.
   The function is typically used in conjunction with the LinePropertiesMatrix function. The function
   is a key part of calculating ImpPTDF.
}
\examples{
 CreateTransmission(g,  EdgeName = "Link", VertexName = "name")
}
\seealso{
\code{\link{LinePropertiesMatrix}} \code{\link{ImpPTDF}}
}
