% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_carefully.R
\name{delete_carefully}
\alias{delete_carefully}
\title{delete_carefully}
\usage{
delete_carefully(g, target, DeleteVect, number, node_name)
}
\arguments{
\item{g}{the graph that is going to be modified}

\item{target}{whether nodes or Edges are being targeted}

\item{DeleteVect}{The vector of names to be deleted. This is taken from the deletion order}

\item{number}{The number of nodes/Edges to be deleted}

\item{node_name}{the node_name of the target e.g "edge_name", or "name"}
}
\description{
This is a helper fucntion for the Fixed and Adaptive strategies. 
It is called by the functions fixed_strategy_attack and adaptive_strategy_attack
It deletes nodes/edges without trying to delete more nodes/edges than there are in the graph. 
It is a bit long and fiddly so is broken out for ease of maintenance.
}
\examples{
delete_carefully(g, target, DeleteVect, number, node_name)
}
\seealso{
\code{\link{fixed_strategy_attack}}, \code{\link{adaptive_strategy_attack}}
}
