% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_limit_models.R
\name{line_limit_volt_pf}
\alias{line_limit_volt_pf}
\alias{line_limit_pf}
\title{Create line limits using a linear model trained on the UK grid}
\usage{
line_limit_volt_pf(g, power_flow, voltage, edge_limit = "edge_limit")

line_limit_pf(g, power_flow, edge_limit = "edge_limit")
}
\arguments{
\item{g}{An igraph object representing a power grid. must have a power flow attribute.}

\item{power_flow}{The node_name of the graph attribute which contains the power flow over each line.}

\item{voltage}{The node_name of the graph attribute which contains the voltage. The attribute should be numeric.}

\item{edge_limit}{The attribute node_name that will be used when setting the line limits.}

\item{node_name}{Line_limits}
}
\value{
The original network with an additional edge attribute that is a numeric line limit created by; Line Limit =  \eqn{a*power_flow + b*voltage + c}
}
\description{
These functions set the line limits of network using a model trained on the UK power grid
}
\details{
The models used here were trained on a network representing a simplified version UK high voltage network similar to that used in the 
    Electricity Ten Year Statement (ETYS) dataset. The model was validated using ten fold cross validation. In testing the Volt PF was a top 
    performer or the best performing model in all tests performed in the paper \href{https://arxiv.org/abs/1907.12848}{Bourne et al. 2019}. 
    The PF model generally performed well but not at the level of the Volt_PF model. 
    The Volt_PF model was trained using three difference voltage levels 132, 275 and 400Kv.
    The relationship between voltage and log10 line limits appeared linear when correcting for power flow. However, we cannot say what will
    happen to the relationship at very high or very low voltage levels. The reccomended order in which artificial line limits should be used
    given available data is as follows.
\enumerate{
  \item line_limit_volt_pf
  \item line_limit_pf
  \item \code{\link{proportional_load}}
}
}
\examples{
#Creat line limits using both methods
test_g_Volt <- line_limit_volt_pf(g, "power_flow", "Voltage", edge_limit = "Line_Limit_VoltPF")
test_g_PF <- line_limit_pf(g, "power_flow", edge_limit = "Line_Limit_PF")
#Compare the resulting line limits.
tibble(VoltPF = get.edge.attribute(test_g_Volt, "Line_Limit"),
    PF =  get.edge.attribute(test_g_PF, "Line_Limit"))
#There can be substantial difference between the two due to the additional information provided by the voltage
    
}
\seealso{
\code{\link{proportional_load}} to create proportionally loaded line limits. 
  \href{https://arxiv.org/abs/1907.12848}{Bourne et al. 2019} describes these models in detail.
}
