% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_centrality.R
\name{electrical_centrality}
\alias{electrical_centrality}
\title{Electrical Edge Centrality}
\usage{
electrical_centrality(g, node_name = "name", generation = "generation",
  demand = "demand", bus_order = "bus_order")
}
\arguments{
\item{g}{An igraph object of a power network.}

\item{node_name}{Is the vertex attribute that contains the node names.}

\item{generation}{The vertex attribute containing the node generation data.}

\item{demand}{The vertec attribute containing the node generation data.}

\item{bus_order}{Only used if no PDF is supplied. The vertex attribute that contains the rank order for slack reference.}
}
\description{
Calculates the electrical edge centrality of the graph
}
\details{
This is an implementation of Wang et al 2011 Physica A, "An electrical betweeness approach for
vulnerability assesment of power grids considering the capacity generators and load
Can only be used on a single component graph, if the graph has multiple components calculate for each component
   individually and add the results together.
}
