% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrical_distance.R
\name{electrical_distance}
\alias{electrical_distance}
\title{Electrical Distance}
\usage{
electrical_distance(imp)
}
\arguments{
\item{imp}{The impedance matrix produced by imp_ptdf}
}
\description{
Calculates the electrical distance between all nodes in the network
}
\details{
Calculates the electrical distance used in Bompard et al 2009, which is defined as \eqn{Z_{ii}+Z_{jj}-2Z_{ij}}. It takes in a matrix and outputs a matrix.
}
\seealso{
\code{\link{imp_ptdf}}
}
