% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectricalCentrality.R
\name{ElectricalCentrality}
\alias{ElectricalCentrality}
\title{Electrical Edge Centrality}
\usage{
ElectricalCentrality(g, Node_name = "name", Generation = "Generation",
  Demand = "Demand", EdgeName = "Link")
}
\arguments{
\item{g}{An igraph object of a power network.}

\item{Node_name}{Is the vertex attribute that contains the node names.}

\item{Generation}{The vertex attribute containing the node generation data.}

\item{Demand}{The vertex attribute containing the node generation data.}

\item{EdgeName}{A character string. The edge attribute that contains the edge names}
}
\description{
Calculates the electrical edge centrality of the graph
}
\details{
This is an implementation of Wang et al 2011 Physica A, "An electrical betweeness approach for
vulnerability assesment of power grids considering the capacity generators and load
Can only be used on a single component graph, if the graph has multiple components calculate for each component
   individually and add the results together.
}
