% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectricalDistance.R
\name{ElectricalDistance}
\alias{ElectricalDistance}
\title{Electrical Distance}
\usage{
ElectricalDistance(Imp)
}
\arguments{
\item{Imp}{The impedance matrix produced by ImpPTDF}
}
\description{
Calculates the electrical distance between all nodes in the network
}
\details{
Calculates the electrical distance used in Bompard et al 2009, which is defined as \eqn{Z_{ii}+Z_{jj}-2Z_{ij}}. It takes in a matrix and outputs a matrix.
}
\seealso{
\code{\link{ImpPTDF}}
}
