% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntropicDegree.R
\name{EntropicDegree}
\alias{EntropicDegree}
\title{Dntropic Degree}
\usage{
EntropicDegree(g, value = "Link.Limit", Scale = FALSE)
}
\arguments{
\item{g}{an igraph object}

\item{value}{The line feature to be used in calculating the entropic degree}

\item{Scale}{Should the values be scaled by multiplying by the log of the node degree}
}
\description{
Calculates the entropic degree of each node in the network using the edge feature of choice.
The default is the flow limit of the line.
}
\details{
This value is implemented as defined by Bompard et al in  "Analysis  of  structural
vulnerabilities  in  power  transmission  grids” doi:  10.1016/j.ijcip.2009.02.002.
}
