% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_attack.R
\name{extend_attack}
\alias{extend_attack}
\title{Extend Attack}
\usage{
extend_attack(network_list, attack_order, min_max_comp = 0,
  total_attack_rounds = 1000)
}
\arguments{
\item{network_list}{The output of attack_the_grid, which is to be extended}

\item{attack_order}{The original attack order used to generate the attack run}

\item{min_max_comp}{The minimum size of the maximum component, default is set to zero}

\item{total_attack_rounds}{The total number of attack rounds to execute in this extension}
}
\description{
Extends a previous attack run with additional attack rounds
}
