% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_multi_attack.R
\name{extend_mullti_attack}
\alias{extend_mullti_attack}
\title{Extend Multi-Attack}
\usage{
extend_mullti_attack(files, deletion_order, min_max_comp = 0,
  total_attack_rounds = 1000)
}
\arguments{
\item{files}{the files in the current working directory to be extended. These files should all have the extension'.rds'}

\item{deletion_order}{The output of multi_attack_order this should be the same as the original dataframe used for the simulations that are being extended}

\item{min_max_comp}{The minimum size of the maximum component, default is set to zero}

\item{total_attack_rounds}{The total number of attack rounds to execute in this extension}
}
\description{
Allows a series of attack runs to be extended.
}
\seealso{
\code{\link{multi_attack_order}} \code{\link{extend_attack}}
}
