% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendAttack.R
\name{ExtendAttack}
\alias{ExtendAttack}
\title{Extend Attack}
\usage{
ExtendAttack(
  NetworkList,
  AttackOrder,
  MinMaxComp = 0,
  TotalAttackRounds = 1000
)
}
\arguments{
\item{NetworkList}{The output of AttackTheGrid, which is to be extended}

\item{AttackOrder}{The original attack order used to generate the attack run}

\item{MinMaxComp}{The minimum size of the maximum component, default is set to zero}

\item{TotalAttackRounds}{The total number of attack rounds to execute in this extension}
}
\description{
Extends a previous attack run with additional attack rounds
}
