% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendMultiAttack.R
\name{ExtendMulltiAttack}
\alias{ExtendMulltiAttack}
\title{Extend Multi-Attack}
\usage{
ExtendMulltiAttack(files, DeleteOrders, MinMaxComp = 0,
  TotalAttackRounds = 1000)
}
\arguments{
\item{files}{the files in the current working directory to be extended. These files should all have the extension'.rds'}

\item{DeleteOrders}{The output of MultiAttackOrder this should be the same as the original dataframe used for the simulations that are being extended}

\item{MinMaxComp}{The minimum size of the maximum component, default is set to zero}

\item{TotalAttackRounds}{The total number of attack rounds to execute in this extension}
}
\description{
Allows a series of attack runs to be extended.
}
\seealso{
\code{\link{MultiAttackOrder}} \code{\link{ExtendAttack}}
}
