% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_attack_stats.R
\name{extract_attack_stats}
\alias{extract_attack_stats}
\title{Extracts Attack stats from saved}
\usage{
extract_attack_stats(root_folder, newfolder_path,
  generation = "net_generation", edge_name = "name",
  power_flow = "power_flow", edge_limit = "edge_limit")
}
\arguments{
\item{root_folder}{:The folder where the attack files are saved}

\item{generation}{The node_name of the variable that stores the net generation data. character string}

\item{edge_name}{A character string. The default is "name".}

\item{power_flow}{A character String. The default is "power_flow".}

\item{edge_limit}{A character string. The default is "edge_limit".}

\item{NewFolderPath}{: The folder where the summary files will be saved to}
}
\description{
Extracts the attack statistics for each network in a folder and combines it into a single file that is also saved.
This function only works on a folder of folders...It needs to be changed
}
