% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_network_stats.R
\name{extract_network_stats}
\alias{extract_network_stats}
\title{Summarise attack Statistics}
\usage{
extract_network_stats(network_list, generation = "net_generation",
  edge_name = "name", power_flow = "power_flow",
  edge_limit = "edge_limit")
}
\arguments{
\item{network_list}{The output of the AttackTheGridFunction}

\item{generation}{The node_name of the variable that stores the net generation data. character string}

\item{edge_name}{A character string. The default is "name".}

\item{power_flow}{A character String. The default is "power_flow".}

\item{edge_limit}{A character string. The default is "edge_limit".}
}
\value{
A dataframe
}
\description{
Sumamrises the attack statistics by extracting the information from
  the list of lists of graphs produced by the attack_the_grid function.
}
\details{
The function summarises the effect of the attack on the network providing statistics on,
Max component size, Total nodes, Total Edges, Power Generated, Grid loading, Giant component fraction,
Blackout size.
}
