% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNetworkStats.R
\name{ExtractNetworkStats}
\alias{ExtractNetworkStats}
\title{Summarise attack Statistics}
\usage{
ExtractNetworkStats(
  NetworkList,
  Generation = "BalencedPower",
  EdgeName = "name",
  PowerFlow = "PowerFlow",
  Link.Limit = "Link.Limit"
)
}
\arguments{
\item{NetworkList}{The output of the AttackTheGridFunction}

\item{Generation}{The name of the variable that stores the net generation data. character string}

\item{EdgeName}{A character string. The default is "name".}

\item{PowerFlow}{A character String. The default is "PowerFlow".}

\item{Link.Limit}{A character string. The default is "Link.Limit".}
}
\value{
A dataframe
}
\description{
Sumamrises the attack statistics by extracting the information from
  the list of lists of graphs produced by the AttackTheGrid function.
}
\details{
The function summarises the effect of the attack on the network providing statistics on,
Max component size, Total Nodes, Total Edges, Power Generated, Grid loading, Giant component fraction,
Blackout size.
}
