% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_strategy_attack.R
\name{fixed_strategy_attack}
\alias{fixed_strategy_attack}
\title{Flexible Removal method}
\usage{
fixed_strategy_attack(g, deletion_order, target = "nodes",
  node_name = "name", number = 1)
}
\arguments{
\item{g}{An igraph object representing a power-grid}

\item{target}{The element that will be attacked. A string, the default is nodes. Can also be Edge.}

\item{node_name}{the variable which holds the edge/node names.}

\item{number}{The total number of targets to calulcate the order for}

\item{DeltionOrder}{The order in which nodes will be targeted for deltion. This is a string vector}
}
\description{
This Removal method pre-calculates the attack order and removes a fixed number of nodes
}
