% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedStrategyAttack.R
\name{FixedStrategyAttack}
\alias{FixedStrategyAttack}
\title{Flexible Removal method}
\usage{
FixedStrategyAttack(
  g,
  DeletionOrder,
  Target = "Nodes",
  Name = "name",
  Number = 1
)
}
\arguments{
\item{g}{An igraph object representing a power-grid}

\item{DeletionOrder}{The order in which nodes will be targeted for deltion. This is a string vector}

\item{Target}{The element that will be attacked. A string, the default is Nodes. Can also be Edge.}

\item{Name}{the variable which holds the edge/node names.}

\item{Number}{The total number of targets to calulcate the order for}
}
\description{
This Removal method pre-calculates the attack order and removes a fixed number of nodes
}
