% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_attack_order.R
\name{generate_attack_order}
\alias{generate_attack_order}
\title{Generate Attack order in multiple simulations}
\usage{
generate_attack_order(sim_order, folder)
}
\arguments{
\item{sim_order}{The dataframe produced by the multi_attack_order}

\item{folder}{the folder that the attack simuations are being saved to}
}
\description{
Takes the output from the 'multi_attack_order' function and produces an attack vector of the next simulation to be run
   The function is used when a large number of simulations are being run and saved to a folder for reuse.
   This function is usally only used as part of \code{\link{save_multi_attacks}}
}
\examples{
generate_attack_order(sim_order, folder)
}
\seealso{
\code{\link{multi_attack_order}}
}
