% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateAttackOrder.R
\name{GenerateAttackOrder}
\alias{GenerateAttackOrder}
\title{Generate Attack order in multiple simulations}
\usage{
GenerateAttackOrder(SimOrder, folder)
}
\arguments{
\item{SimOrder}{The dataframe produced by the MultiAttackOrder}

\item{foler}{the folder that the attack simuations are being saved to}
}
\description{
Takes the output from the 'MultiAttackOrder' function and produces an attack vector of the next simulation to be run
The function is used when a large number of simulations are being run and saved to a folder for reuse.
}
\examples{
GenerateAttackOrder(SimOrder, folder)
}
\seealso{
MultiAttackOrder
}
