% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_ptdf.R
\name{imp_ptdf}
\alias{imp_ptdf}
\title{Create Impedance matrix and ptdf}
\usage{
imp_ptdf(g, slack_ref, edge_name = "edge_name", node_name = "name")
}
\arguments{
\item{g}{An igraph object representing a power-grid}

\item{slack_ref}{The slack node for the power-grid, A character vector}

\item{edge_name}{the variable that holds the edge names, a character string.}

\item{node_name}{the variable that holds the names of the nodes, to identify the slack ref. a character string}
}
\description{
This is a wrapper for the more interesting parts of the electrical building blocks
  It creates both matrices as they almost always both need to be created and they take time so producing
  together is more efficient than producing each one on its own.
}
\examples{
g_2 <-make_ego_graph(g, 2, "AXMI")[[1]]
imp_ptdf(g, slack_ref)
}
