% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_loading.R
\name{line_loading}
\alias{line_loading}
\title{Line loading
Calculates the fraction of maximum flow in each edge across a power-grid}
\usage{
line_loading(g, edge_name = "name", power_flow = "power_flow",
  edge_limit = "edge_limit")
}
\arguments{
\item{g}{An igraph object representing a powergrid.}

\item{edge_name}{A character string. The default is "name".}

\item{power_flow}{A character String. The default is "power_flow".}

\item{link_limit}{A character string. The default is "link_limit".}
}
\value{
A dataframe
}
\description{
Line loading is used as a way of describing how close a power-grid or power line is to tripping.
Line loading is the inverse of tolerance (\eqn{\alpha}).
}
\examples{
load_level(g)
}
