% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_properties_matrix.R
\name{line_properties_matrix}
\alias{line_properties_matrix}
\title{Create the line Properties matrix}
\usage{
line_properties_matrix(g, edge_name = "edge_name", weight = "y")
}
\arguments{
\item{g}{Igraph object which contains an attribute naming the edges and another that provides the edge weight}

\item{edge_name}{A character string of the edge attribute that holds edge names. Default is "edge_name"}

\item{weight}{A character string of the edge.attribute for edge weights. Default is "Y}
}
\value{
A matrix, with names rows and columns
}
\description{
Creates the line properties matrix
}
\details{
This function creates a line properties matrix that can be used with the output of the Transmission matrix function.
 The function takes an igraph object of a power-grid and returns diagonal matrix. Where the diagonal is the edgeweight of
 each edge.
}
\examples{
This example doesn't show in the documentatio, I don't know why
g_2 <-make_ego_graph(g, 2, "AXMI")[[1]]
plot(g_2)
line_properties_matrix(g_2)
}
