% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinePropertiesMatrix.R
\name{LinePropertiesMatrix}
\alias{LinePropertiesMatrix}
\title{Create the line Properties matrix}
\usage{
LinePropertiesMatrix(g, Edgename = "Link", Weight = "Y")
}
\arguments{
\item{g}{Igraph object which contains an attribute naming the edges and another that provides the edge weight}

\item{Edgename}{A character string of the edge attribute that holds edge names. Default is "Link"}

\item{Weight}{A character string of the edge.attribute for edge weights. Default is "Y}
}
\value{
A matrix, with names rows and columns
}
\description{
Creates the line properties matrix representing
}
\details{
This function creates a line properties matrix that can be used with the output of the Transmission matrix function.
 The function takes an igraph object of a power-grid and returns diagonal matrix. Where the diagonal is the edgeweight of
 each edge.
}
