% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_level.R
\name{load_level}
\alias{load_level}
\title{Get load level of power-grid
Calculates the fraction of maximum flow in each edge across a power-grid}
\usage{
load_level(g, edge_name = "name", power_flow = "power_flow",
  edge_limit = "edge_limit")
}
\arguments{
\item{g}{An igraph object representing a powergrid.}

\item{edge_name}{A character string. The default is "name".}

\item{power_flow}{A character String. The default is "power_flow".}

\item{edge_limit}{A character string. The default is "edge_limit".}
}
\value{
A dataframe
}
\description{
Load level is used as a way of describing how close a power-grid or power line is to tripping.
Load level is the inverse of tolerance (\eqn{\alpha}).
}
\examples{
load_level(g)
}
