% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadLevel.R
\name{LoadLevel}
\alias{LoadLevel}
\title{Get load level of power-grid
Calculates the fraction of maximum flow in each edge across a power-grid}
\usage{
LoadLevel(
  g,
  EdgeName = "name",
  PowerFlow = "PowerFlow",
  Link.Limit = "Link.Limit"
)
}
\arguments{
\item{g}{An igraph object representing a powergrid.}

\item{PowerFlow}{A character String. The default is "PowerFlow".}

\item{Link.Limit}{A character string. The default is "Link.Limit".}

\item{Edgename}{A character string. The default is "name".}
}
\value{
A dataframe
}
\description{
Load level is used as a way of describing how close a power-grid or power line is to tripping.
Load level is the inverse of tolerance (\eqn{\alpha}).
}
\examples{
LoadLevel(g)
}
