% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_or_zero.R
\name{missing_or_zero}
\alias{missing_or_zero}
\title{Missing or Zero replacement}
\usage{
missing_or_zero(data, type)
}
\arguments{
\item{data}{A vector}

\item{type}{A character string. either max or min}
}
\description{
I have no idea if function is used or needs to be deleted.
}
\examples{
missing_or_zero(data, type)
}
