% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_attack_order.R
\name{multi_attack_order}
\alias{multi_attack_order}
\title{Generate a dataframe of deletion orders}
\usage{
multi_attack_order(g, target = "nodes", sims, node_name = "name")
}
\arguments{
\item{g}{an igraph object.}

\item{target}{whether Edges or nodes will be attacked}

\item{sims}{the number of attack simulations to be performed}

\item{node_name}{The attribute node_name of the target}
}
\description{
This function takes an igraph object and returns a dataframe where each row represents a node/edge
   deletion order. It is used for simulating a large amount of random attacks in a reproducible way.
}
\examples{
ExampleTargets <- multi_attack_order(g, 100)
}
