% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiAttackOrder.R
\name{MultiAttackOrder}
\alias{MultiAttackOrder}
\title{Generate a dataframe of deletion orders}
\usage{
MultiAttackOrder(g, Sims)
}
\arguments{
\item{g}{an igraph object.}

\item{Sims}{the number of attack simulations to be performed}
}
\description{
This function takes an igrap object and returns a dataframe where each row represents a node/edge
   deletion order. It is used for simulating a large amount of attacks in a reproducible way.
   Currently the function only generates random attacks.
}
\examples{
ExampleTargets <- MultiAttackOrder(g, 100)
}
