% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_ability.R
\name{net_ability}
\alias{net_ability}
\title{Net-ability}
\usage{
net_ability(g, g_orig = NULL)
}
\arguments{
\item{g}{The graph for which net-ability will be calculated}

\item{g_orig}{The original graph structure. The defualt is the g}
}
\description{
This function calculates the net-ability of a graph.
}
\details{
The algorithm was implemented from "power grids vulnerability a complex networks approach" Arianos 2009 and
"Analysis of structural vulnerabilities in power transmission grids" Bompard 2009.
\code{g_orig} is an important vairable if your network has already been attacked. If generation or demand nodes are removed
 from the graph due to failure they will not be considered in the net-ability potentially increasing new graphs score as
 the total number of node pairs will be smaller in the newgraph. Because of this the number of node pairs always has to be
 calculated from the original graph
}
\examples{
NetAbScore <-net_ability(g)
}
