% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_co_failure.R
\name{network_co_failure}
\alias{network_co_failure}
\title{Network co-failure}
\usage{
network_co_failure(collapse_list)
}
\arguments{
\item{collapse_list}{A list. This is the output of the attack_the_grid function}
}
\description{
This function finds the co-failure and failure mode of the node and edges in the attack on the network
}
\details{
This function produces a 6 element list. The node co failure matrix, the edge co failure matrix, the co-failure matrices for both
edge and node but excluding targeting. The final two elements are dataframes of the failure mode for the edges and nodes
}
