% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_stats.R
\name{network_stats}
\alias{network_stats}
\title{network_stats
Creates summary statistics of the network}
\usage{
network_stats(g, colname = "value")
}
\arguments{
\item{g}{An igraph object}

\item{colname}{A character string. The desired node_name of the output statistic column. the default is "value"}
}
\description{
The function takes a graph and returns a set of common summary graph statistics to understand the graph topology.
These statistics are, number of node, number of edges, number of components, Betweeness, Mean Degree, Assortivity, Clustering, and Mean Distance.
}
