% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_attack_simulation.R
\name{next_attack_simulation}
\alias{next_attack_simulation}
\title{Find the next attack simulation to be run}
\usage{
next_attack_simulation(sim_order, folder)
}
\arguments{
\item{sim_order}{The output dataframe of multi_attack_order}

\item{folder}{The folder where the attack simulations are saved}
}
\description{
When multiple attack simulations are being performed the operation can be interrupted.
  This function searches the folder where the simulations are being stored and returns
  the next attack simulation. This function is mostly only used as part of the multi_attack function
}
\examples{
  next_attack_simulation(sim_order, folder)

}
\seealso{
\code{\link{multi_attack_order}}, \code{\link{multi_attack}}
}
