% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NextAttackSimulation.R
\name{NextAttackSimulation}
\alias{NextAttackSimulation}
\title{Find the next attack simulation to be run}
\usage{
NextAttackSimulation(SimOrder, folder)
}
\arguments{
\item{SimOrder}{The output dataframe of MultiAttackOrder}

\item{folder}{The folder where the attack simulations are saved}
}
\description{
When multiple attack simulations are being performed the operation can be interrupted.
  This function searches the folder where the simulations are being stored and returns
  the next attack simulation.
}
\seealso{
MultiAttackOrder
}
