% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_pairs.R
\name{node_pairs}
\alias{node_pairs}
\title{Node pairs}
\usage{
node_pairs(g, node_name = "name", generation = "generation",
  demand = "demand", ptdf = NULL, bus_order = "bus_order")
}
\arguments{
\item{g}{An igraph object of a power network.}

\item{node_name}{Is the vertex attribute that contains the node names.}

\item{generation}{The vertex attribute containing the node generation data.}

\item{demand}{The vertec attribute containing the node generation data.}

\item{ptdf}{An optional arguement. The PDTF matrix as produced by imp_ptdf. default is NULL. If not ptdf is supplied one is calculated.}

\item{bus_order}{Only used if no PDF is supplied. The vertex attribute that contains the rank order for slack reference.}
}
\description{
This function finds every pair of nodes in the network. It takes in a igraph power network and outputs a dataframe.
The dataframe contains the node pair names, their node ID number type of node generation/demand/Transmission, the Net generation of
the node, and wether the node pair is generation and demand
}
\seealso{
\code{\link{slack_ref_func}}
}
