% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodePairs.R
\name{NodePairs}
\alias{NodePairs}
\title{Node pairs}
\usage{
NodePairs(g, Node_name = "name", Generation = "Generation",
  Demand = "Demand", PTDF = NULL, Bus_order = "Bus.Order")
}
\arguments{
\item{g}{An igraph object of a power network.}

\item{Node_name}{Is the vertex attribute that contains the node names.}

\item{Generation}{The vertex attribute containing the node generation data.}

\item{Demand}{The vertec attribute containing the node generation data.}

\item{PTDF}{An optional arguement. The PDTF matrix as produced by ImpPTDF. default is NULL. If not PTDF is supplied one is calculated.}

\item{Bus_order}{Only used if no PDF is supplied. The vertex attribute that contains the rank order for slack reference.}
}
\description{
This function finds every pair of nodes in the network. It takes in a igraph power network and outputs a dataframe.
The dataframe contains the node pair names, their node ID number type of node Generation/Demand/Transmission, the Net generation of
the node, and wether the node pair is generation and demand
}
\seealso{
\code{\link{SlackRefFunc}}
}
