% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_targeted.R
\name{nodes_targeted}
\alias{nodes_targeted}
\title{Nodes Targeted}
\usage{
nodes_targeted(network_list)
}
\arguments{
\item{network_list}{a list of lists of networks. The output of Attack the grid}
}
\description{
Gets the node_name of the deleted target node/edge from the named attribute in each attack list.
}
\examples{
nodes_targeted(network_list)
}
\seealso{
\code{\link{attack_the_grid}}
}
