% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_flow.R
\name{power_flow}
\alias{power_flow}
\title{Calculate DC power flow}
\usage{
power_flow(g, slack_ref, edge_name = "edge_name", node_name = "name",
  net_generation = "net_generation")
}
\arguments{
\item{g}{An igraph object representing the power grid}

\item{slack_ref}{A character strong. The node to remove from the calculations to prevent the matrix being singular}

\item{edge_name}{The variable that holds the edge names, a character string.}

\item{node_name}{The variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{net_generation}{The node_name that the net generation data for each node is held in}
}
\description{
Calculates the power_flow from a graph that contains the following attributes,
named edges, edgeweights, a balanced power generation and demand column. Writes an attribute called Powerflow 
(existing attributeof the same node_name will be overwritten). 
The function outputs a a graph with the correct power flow values
}
\examples{
power_flow(g Slackref)
}
\keyword{flow}
\keyword{power}
