% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerFlow.R
\name{PowerFlow}
\alias{PowerFlow}
\title{Calculate DC power flow}
\usage{
PowerFlow(g, SlackRef)
}
\arguments{
\item{g}{an igraph object}

\item{SlackRef}{the node to remove from the calculations to prevent the matrix being singular}
}
\description{
TCalculates the PowerFlow from a graph that contains the following attributes,
named edges, edgeweights, a balanced power generation and demand column, Powerflow (will be overwritten)
}
\examples{
PowerFlow(g Slackref)
}
\keyword{flow}
\keyword{power}
