% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_attack.R
\name{random_attack}
\alias{random_attack}
\title{Generate a random attack order}
\usage{
random_attack(g, target = "nodes", number = 1, node_name = "name")
}
\arguments{
\item{g}{An igraph object representing a power-grid.}

\item{target}{The target of the attack. A string either "nodes" or "Edges".}

\item{number}{The number of targets to generate. An integer the default is set to 1.}

\item{node_name}{The attribute node_name of the target}
}
\description{
The most simple attack strategy. Targets either nodes or edges.
}
\examples{
random_attack(g, number = V(g))
}
