% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomAttack.R
\name{RandomAttack}
\alias{RandomAttack}
\title{Generate a random attack order}
\usage{
RandomAttack(g, Target = "Nodes", Number = 1)
}
\arguments{
\item{g}{An igraph object representing a power-grid.}

\item{Target}{The target of the attack. A string either "Nodes" or "Edges".}

\item{Number}{The number of targets to generate. An integer the default is set to 1.}
}
\description{
The most simple attack strategy. Targets either nodes or edges.
}
