% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dead_ends.R
\name{remove_dead_ends}
\alias{remove_dead_ends}
\title{Remove Dead Ends}
\usage{
remove_dead_ends(g, iteration = 1, type = "transfer")
}
\arguments{
\item{g}{An igraph object representing a power network}

\item{iteration}{the number of times the function has recursively removed nodes. Default start is 1}
}
\description{
Removes Transfernodes of Degree 1 from the network. This prevents the network being padded with nodes that don't do anything
}
