% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_symbols.R
\name{remove_symbols}
\alias{remove_symbols}
\title{remove_symbols}
\usage{
remove_symbols(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
probably shouldn't be in this package. I need to chek the dependencies before removing
}
