% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveDeadEnds.R
\name{RemoveDeadEnds}
\alias{RemoveDeadEnds}
\title{Remove Dead Ends}
\usage{
RemoveDeadEnds(g, iteration = 1)
}
\arguments{
\item{g}{An igraph object representing a power network}

\item{iteration}{}
}
\description{
Removes Transfernodes of Degree 1 from the network. This prevents the network being padded with nodes that don't do anything
}
