% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_removed.R
\name{round_removed}
\alias{round_removed}
\title{Find Deletion order}
\usage{
round_removed(network_list, type = "edge")
}
\arguments{
\item{network_list}{The output of attack_the_grid, a list of lists containing igraphs.}

\item{type}{Whether Edge or node order will be found. The default is Edge, any other value will mean nodes are used.}
}
\description{
Attack round that each edge/node was deleted is  extracted from the network List of lists produced by attack_the_grid.
Whether the node/edge was the targeted node/edge is also marked. This currently only works for nodes not Edges
}
\details{
If not all edges/nodes were deleted during attack_the_grid, these edges will be filled with NA
}
\seealso{
\code{\link[PowerGridNetworking]{attack_the_grid}}
}
