% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_type_removed.R
\name{round_type_removed}
\alias{round_type_removed}
\title{Round Removed and type of Removal}
\usage{
round_type_removed(attack_series)
}
\arguments{
\item{attack_series}{The list of lists produced by attack the grid}
}
\description{
This function finds the round removed of all nodes and Edges in a network as well as whether the removal type was Targeted/Overloaded/Islanded
}
\details{
The function accepts the list of lists produced by the function attack the grid and outputs a
    list containing two dataframes for nodes and Edges respectively.
}
