% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundRemoved.R
\name{RoundRemoved}
\alias{RoundRemoved}
\title{Find Edge Deletion order}
\usage{
RoundRemoved(ListofLists)
}
\arguments{
\item{ListofLists}{The output of AttackTheGrid, a list of lists containing igraphs.}
}
\description{
Attack round that each edge was deleted is  extracted from the network List of lists produced by AttackTheGrid
}
\details{
If not all edges were deleted during AttackTheGrid, these edges will be filled with NA
}
\seealso{
\code{\link[PowerGridNetworking]{AttackTheGrid}}
}
