% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundTypeRemoved.R
\name{RoundTypeRemoved}
\alias{RoundTypeRemoved}
\title{Round Removed and Type of Removal}
\usage{
RoundTypeRemoved(AttackSeries)
}
\arguments{
\item{AttackSeries}{The list of lists produced by attack the grid}
}
\description{
This function finds the round removed of all Nodes and Edges in a network as well as whether the removal type was Targeted/Overloaded/Islanded
}
\details{
The function accepts the list of lists produced by the function attack the grid and outputs a
    list containing two dataframes for Nodes and Edges respectively.
}
