% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_multi_attacks.R
\name{save_multi_attacks}
\alias{save_multi_attacks}
\title{Save attack simulations to folder}
\usage{
save_multi_attacks(g, attack_vectors, folder, min_max_comp = 0,
  total_attack_rounds = 10, cascade_mode = FALSE, demand = "demand",
  generation = "generation", edge_name = "edge_name",
  node_name = "name", net_generation = "net_generation",
  target = "nodes")
}
\arguments{
\item{g}{An igraph object representing a power grid.}

\item{attack_vectors}{A dataframe. This is the output of the multi_attack_order function. If you expect.
to interrupt the function this dataframe should be saved or the random seed used to generate the dataset
recorded.}

\item{folder}{the folder data will be saved to.}

\item{min_max_comp}{See attack_the_grid. Default is 0}

\item{total_attack_rounds}{See attack_the_grid. Default is 10}

\item{cascade_mode}{See attack_the_grid. Default is set to FALSE}

\item{demand}{the node_name of the node Load variable. A character string.}

\item{generation}{the node_name of the node generation variable. A character string.}

\item{edge_name}{the variable that holds the edge names, a character string.}

\item{node_name}{the variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{net_generation}{the node_name that the net generation data for each node is held in}

\item{target}{whether nodes or edges are being attacked}
}
\description{
When creating many attack simulations it can be conveniant to save the resulting graphs to a folder
   so that re-runs are not required if further analysis is needed. This function saves the
   results of attack_the_grid, and also allows the process to be interrupted.
   When the process is started again after and interuption the function finds the last
   save attack and starts the simulation process from there.
}
\examples{
save_multi_attacks(g, attack_vectors, folder, cascade_mode = F)
}
\seealso{
\code{\link{attack_the_grid}}, \code{\link{multi_attack_order}}
}
\keyword{multi-attack,}
