% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveMultiAttacks.R
\name{SaveMultiAttacks}
\alias{SaveMultiAttacks}
\title{Save attack simulations to folder}
\usage{
SaveMultiAttacks(g, AttackVectors, folder, MinMaxComp = 0,
  TotalAttackRounds = 10, CascadeMode = FALSE)
}
\arguments{
\item{g}{An igraph object representing a power grid.}

\item{AttackVectors}{A dataframe. This is the output of the MultiAttackOrder function. If you expect.
to interrupt the function this dataframe should be saved or the random seed used to generate the dataset
recorded.}

\item{folder}{the folder data will be saved to.}

\item{MinMaxComp}{See AttackTheGrid. Default is 0}

\item{TotalAttackRounds}{See AttackTheGrid. Default is 10}

\item{CascadeMode}{See AttackTheGrid. Default is set to FALSE}
}
\description{
When creating many attack simulations it can be conveniant to save the resulting graphs to a folder
   so that re-runs are not required if further analysis is needed. This function saves the
   results of AttackTheGrid, and also allows the process to be interrupted.
   When the process is started again after and interuption the function finds the last
   save attack and starts the simulation process from there.
}
\seealso{
\code{\link{AttackTheGrid}}, \code{\link{MultiAttackOrder}}
}
\keyword{multi-attack,}
