% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_ref.R
\name{slack_ref_func}
\alias{slack_ref_func}
\title{slack_ref}
\usage{
slack_ref_func(g, node_name = "name", generation = "generation")
}
\arguments{
\item{g}{The igraph object representing the power grid}

\item{node_name}{The vertex attribute containing node names}

\item{generation}{The vertex attribute that contains the generation data.}
}
\description{
Finds the Slack node in each component of the power grid. A graph is put in and a data frame giving
the slack reference for each component is returned. The Slack bus is the largest generator in that component
}
