% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_multi_attack.R
\name{summarise_multi_attack}
\alias{summarise_multi_attack}
\title{Summarise Multi-Attack}
\usage{
summarise_multi_attack(df, grouping_var = alpha)
}
\arguments{
\item{df}{A dataframe that is produced by the extract_network_stats function.}

\item{grouping_var}{The bare unquoted node_name of the grouping variable, the default is alpha}
}
\description{
This function provides a consice way to summarise the output of the extract_attack_stats function, especially when said function
has been used over a lot of attacks. This is most common when using "SaveMultiAttacks".
}
\examples{
summarise_multi_attack(df)
}
