% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummariseMultiAttack.R
\name{SummariseMultiAttack}
\alias{SummariseMultiAttack}
\title{Summarise Multi-Attack}
\usage{
SummariseMultiAttack(df, groupingvar = alpha)
}
\arguments{
\item{df}{A dataframe that is produced by the ExtractStats function.}

\item{groupingvar}{The bare unquoted name of the grouping variable, the default is alpha}
}
\description{
This function provides a consice way to summarise the output of the ExtractAttackStats function, especially when said function
has been used over a lot of attacks. This is most common when using "SaveMultiAttacks".
}
