% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_capacity.R
\name{transfer_capacity}
\alias{transfer_capacity}
\title{Transfer capacity}
\usage{
transfer_capacity(g, ptdf)
}
\arguments{
\item{g}{An igraph object}

\item{ptdf}{a matrix. This is the output of the imp_ptdf function}
}
\description{
This function is an intermediate step in the calculation of net-ability
}
\examples{
transfer_capacity(g, ptdf)
}
\seealso{
\code{\link{net_ability}}
}
