% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_2.R
\name{union_2}
\alias{union_2}
\title{Merge two graphs}
\usage{
union_2(g_1, g_2)
}
\arguments{
\item{g_1}{Igraph object}

\item{g_2}{Igraph object}
}
\value{
A new graph object
}
\description{
Merges two graphs maintaining whilst thier attributes.
}
\details{
This function is an upgrade to the \pkg{igraph} function "union".
   It creates the union of two networks, and merges attributes with the same node_name.
   In the original union the graph attributes were lost.
   In the case where there are overlapping nodes the attributes of g_1 take precedence
}
\examples{
union_2(g_1,g_2)
Reduce(union_2, list(g_1,g_2,g3))

}
\seealso{
\code{\link[igraph]{union.igraph}}
}
