% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveStrategyAttack.R
\name{AdaptiveStrategyAttack}
\alias{AdaptiveStrategyAttack}
\title{Make Strategy Adaptive}
\usage{
AdaptiveStrategyAttack(g, AttackStrategy, Target = "Nodes", Number = 1)
}
\arguments{
\item{g}{The network which is being attacked}

\item{Number}{The Number of targets to generate}

\item{Nodes}{The Number of Nodes to Remove}

\item{seed}{The random number generator seed to use. This is optional}
}
\description{
This implements an attack strategy to be adaptive.
    This means that it will be re-evaluated every
    round to choose the most effective point of attack
    according to that metric
}
\examples{
AdaptiveStrategyAttack <- function(g, AttackStrategy, Target = "Nodes", Number = 1)
}
\keyword{regime}
\keyword{removal}
