% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_the_grid.R
\name{attack_the_grid}
\alias{attack_the_grid}
\title{Initiate power-grid attack simulation}
\usage{
attack_the_grid(g, AttackStrategy, g0 = NULL, TotalAttackRounds = 1000,
  CascadeMode = TRUE, Demand = "demand", Generation = "generation",
  EdgeName = "edge_name", VertexName = "name",
  Net_generation = "net_generation", power_flow = "power_flow",
  edge_capacity = "edge_capacity", target = "nodes")
}
\arguments{
\item{AttackStrategy}{A function that calculates which node to delete the function is is in "quo" form and embedded in an
attack type.}

\item{g0}{The grid that will be used to test the largest component against if NULL it uses the given network.}

\item{TotalAttackRounds}{The maximum number of nodes to be removed before the process stops.}

\item{CascadeMode}{Whether the power flow equations will be used to check line-overloading or not.}

\item{Demand}{the name of the node Load variable. A character string.}

\item{Generation}{the name of the node generation variable. A character string.}

\item{EdgeName}{the variable that holds the edge names, a character string.}

\item{VertexName}{the variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{Net_generation}{the name that the net generation data for each node is held in}

\item{power_flow}{A character string. This value indicates the name of the edge attribute that holds power flow, the default is "PowerFlow"}

\item{edge_capacity}{A character string. This value indicates the name of the edge attribute that holds the edge limit, the default is "Link.Limit"}

\item{target}{A character string. This value defines the columns of the returned it must match the target element of the attack strategy or an
error will be thrown if the number of columns of the matrix is exceeded. modularising the function call to attack strategy
would solve this, but it is a low priority.}

\item{NetworkList}{A list of lists where each element of the sub list is an igraph object, the first time it is used the
the network list is simply list(list(g)).}

\item{MinMaxComp}{The minimum size of the maximum component, as a fraction, for the process to continue, the default is set
to 0.0 complete collapse.}
}
\description{
This function simulates an attack on the power grid using the parameter settings you choose
   the outut of the function is a nested list of igraph objects.
}
\examples{
AttackTheGrid(NetworkList, AttackStrategy, SubstationData, EdgeData, g0 = NULL)
Out <- AttackTheGrid(NetworkList,
AttackStrategy,
g0 = NULL,
TotalAttackRounds=100,
CascadeMode = TRUE,
CumulativeAttacks = NULL)
}
