% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttackTheGrid.R
\name{AttackTheGrid}
\alias{AttackTheGrid}
\title{Initiate power-grid attack simulation}
\usage{
AttackTheGrid(NetworkList, AttackStrategy, referenceGrid = NULL,
  MinMaxComp = 0.8, TotalAttackRounds = 1000, CascadeMode = TRUE,
  CumulativeAttacks = NULL, Demand = "Demand",
  Generation = "Generation", EdgeName = "Link", VertexName = "name",
  Net_generation = "BalencedPower")
}
\arguments{
\item{NetworkList}{A list of lists where each element of the sub list is an igraph object, the first time it is used the
the network list is simply list(list(g)).}

\item{AttackStrategy}{A function that calculates which node to delete the function is is in "quo" form and embedded in an
attack type.}

\item{referenceGrid}{The grid that will be used to test the largest component against if NULL it uses the given network.}

\item{MinMaxComp}{The minimum size of the maximum component, as a fraction, for the process to continue, the default is set
to 0.8.}

\item{TotalAttackRounds}{The maximum number of nodes to be removed before the process stops.}

\item{CascadeMode}{Whether the power flow equations will be used to check line-overloading or not.}

\item{CumulativeAttacks}{The total number of attacks that have taken place so far.}

\item{Demand}{the name of the node Load variable. A character string.}

\item{Generation}{the name of the node generation variable. A character string.}

\item{EdgeName}{the variable that holds the edge names, a character string.}

\item{VertexName}{the variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{Net_generation}{the name that the net generation data for each node is held in}
}
\description{
This function simulates an attack on the power grid using the parameter settings you choose
   the outut of the function is a nested list of igraph objects.
}
\examples{
AttackTheGrid(NetworkList, AttackStrategy, SubstationData, EdgeData, referenceGrid = NULL, MinMaxComp = 0.8)
Out <- AttackTheGrid(NetworkList,
AttackStrategy,
referenceGrid = NULL,
MinMaxComp = 0.8,
TotalAttackRounds=100,
CascadeMode = TRUE,
CumulativeAttacks = NULL)
}
