% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcOverLimit.R
\name{CalcOverLimit}
\alias{CalcOverLimit}
\title{Find which edges are over the limit}
\usage{
CalcOverLimit(g)
}
\arguments{
\item{g}{An igraph object that represents a power-grid}
}
\description{
This function checks to see which lines are over limit and removes them and re-calculates power flow in the network
}
\details{
The function uses a node attribute "Bus.Order" to identify the slack bus. This may be changed in later versions
}
