% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{cascade}
\alias{cascade}
\title{Calculate the effects of a cascade}
\usage{
cascade(network_list, iteration = 0, stop_cascade = Inf, g_0 = NULL,
  generation = "generation", demand = "demand",
  edge_name = "edge_name", node_name = "name",
  net_generation = "net_generation")
}
\arguments{
\item{network_list}{A list of power-grid networks.}

\item{iteration}{The current iteration number}

\item{stop_cascade}{The number of iterations when cascade will be forced to terminate. An integer, default is set to infinity.}

\item{generation}{the node_name of the node generation variable. A character string.}

\item{demand}{the node_name of the node Load variable. A character string.}

\item{edge_name}{the variable that holds the edge names, a character string.}

\item{node_name}{the variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{net_generation}{the node_name that the net generation data for each node is held in}
}
\description{
Iterates through the network removing edges until no edges are over the line limit and the network has stabilised.
}
\details{
This is a recursive function that calls itself repeatedly until the cascade has come to a stop and the network has stabilised.
  The function produces a list of graphs showing each state of the cascade. This function is mostly used as part of the
  attack_the_grid function.
}
\seealso{
\code{\link{attack_the_grid}}
}
