% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components_differ.R
\name{components_differ}
\alias{components_differ}
\title{Find components of a graph that have changed
Find components of a graph that have changed}
\usage{
components_differ(g, g_0, edge_name = "name")
}
\arguments{
\item{g}{An igraph object. The current graph}

\item{g_0}{An igraph object. the previous graph}

\item{edge_name}{a character string. The graph attribute describing the edge node_name, default is set to 'node_name'}
}
\description{
Using this should be able to cut down on the amount of calculation done on the power flow equations
This function checks to see which subcomponents have changed in the graph and which are the same.
It does this by comparing edges. This is because a component may have had an edge removed but still be intact.
However, the powerflow will still need to be re-calculated.
The function outputs a logical value
}
