% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectricalCentrality.R
\name{ElectricalCentrality}
\alias{ElectricalCentrality}
\title{Electrical Edge Centrality}
\usage{
ElectricalCentrality(g, Node_name = "name", Generation = "Generation",
  Demand = "Demand", Bus_order = "Bus.Order")
}
\arguments{
\item{g}{An igraph object of a power network.}

\item{Node_name}{Is the vertex attribute that contains the node names.}

\item{Generation}{The vertex attribute containing the node generation data.}

\item{Demand}{The vertec attribute containing the node generation data.}

\item{Bus_order}{Only used if no PDF is supplied. The vertex attribute that contains the rank order for slack reference.}
}
\description{
Calculates the electrical edge centrality of the graph
}
