% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_network_stats.R
\name{extract_network_stats}
\alias{extract_network_stats}
\title{Extract Network stats}
\usage{
extract_network_stats(data_list)
}
\arguments{
\item{data_list}{A list. the output of attack_the_grid}
}
\description{
This function extracts the attack statistics from the output of attack_the_grid
}
\details{
The function is about 55 times faster than ExtractNetworkStats
}
