% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractAttackStats.R
\name{ExtractAttackStats}
\alias{ExtractAttackStats}
\title{Extracts Attack stats from saved}
\usage{
ExtractAttackStats(RootFolder, NewfolderPath)
}
\arguments{
\item{RootFolder}{:The folder where the attack files are saved}

\item{NewFolderPath}{: The folder where the summary files will be saved to}
}
\description{
Extracts the attack statistics for each network in a folder and combines it into a single file that is also saved
}
