% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractNetworkStats.R
\name{ExtractNetworkStats}
\alias{ExtractNetworkStats}
\title{Summarise attack Statistics}
\usage{
ExtractNetworkStats(NetworkList)
}
\arguments{
\item{NetworkList}{The output of the AttackTheGridFunction}
}
\value{
A dataframe
}
\description{
Sumamrises the attack statistics by extracting the information from
  the list of lists of graphs produced by the AttackTheGrid function.
}
\details{
The function summarises the effect of the attack on the network providing statistics on,
Max component size, Total Nodes, Total Edges, Power Generated, Grid loading, Giant component fraction,
Blackout size.
}
