% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpPTDF.R
\name{ImpPTDF}
\alias{ImpPTDF}
\title{Create Impedance matrix and PTDF}
\usage{
ImpPTDF(g, SlackRef)
}
\arguments{
\item{g}{An igraph object representing a power-grid}

\item{SlackRef}{The slack node for the power-grid, A character vector}
}
\description{
This is a wrapper for the more interesting parts of the electrical building blocks
  It creates both matrices as they almost always both need to be created and they take time so producing
  together is more efficient than producing each one on its own.
}
