% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerFlow.R
\name{PowerFlow}
\alias{PowerFlow}
\title{Calculate DC power flow}
\usage{
PowerFlow(g, AZero, LineProperties, EdgeName = "Link",
  VertexName = "name", Net_generation = "BalencedPower",
  power_flow = "PowerFlow")
}
\arguments{
\item{g}{An igraph object representing the power grid}

\item{AZero}{A numeric matrix The transmission matrix of the original network}

\item{LineProperties}{A numeric matrix. a diagonal matrix of the Y characteristic of the network}

\item{EdgeName}{The variable that holds the edge names, a character string.}

\item{VertexName}{The variable that holds the names of the nodes, to identify the slack ref. a character string}

\item{Net_generation}{The name that the net generation data for each node is held in}

\item{power_flow}{A character string. The name of the edge attribute that will hold the power flow information}
}
\description{
Calculates the PowerFlow from a graph that contains the following attributes,
named edges, edgeweights, a balanced power generation and demand column. Writes a new edge attribute of power flow
(existing attribute of the same name will be overwritten). 
The function outputs a a graph with the correct power flow values
}
\examples{
PowerFlow(g Slackref)
}
\keyword{flow}
\keyword{power}
