% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportional_Load.R
\name{Proportional_Load}
\alias{Proportional_Load}
\title{Create proportional loading line limits}
\usage{
Proportional_Load(g, alpha, PowerFlow = "PowerFlow",
  Link.Limit = "Link.Limit")
}
\arguments{
\item{g}{An igraph object representing a power grid. must have a power flow attribute.}

\item{alpha}{The system tolerance. A number larger than 1.}

\item{PowerFlow}{The name of the graph attribute which contains the power flow over each line.}

\item{Link.Limit}{The attribute name that will be used when setting the line limits.}
}
\value{
The original network with an additional edge attribute that is a numeric line limit created by; Line Limit =  \eqn{\alpha*PowerFlow}
}
\description{
This function sets the line limits to the tolerance value \eqn{\alpha} chosen by the user.
}
\details{
Proportional loading should only be used when there are no other options available as it can
   be challenging to set the value of \eqn{\alpha} close to the system mean.
   \href{https://doi.org/10.1063/1.5115493}{Bourne et al. 2019} find that values of \eqn{\alpha} that
   deviate substantially from the actual value of \eqn{\alpha} given that load profile will not represent
   the behaviour of the power grid under attack or collapse.
}
\examples{
Proportional_Load(g, alpha = 1.05)
}
