% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportional_Load.R
\name{Proportional_Load}
\alias{Proportional_Load}
\title{Create proportional loading line limits}
\usage{
Proportional_Load(g, alpha, PowerFlow = "PowerFlow",
  Link.Limit = "Link.Limit")
}
\arguments{
\item{g}{An igraph object representing a power grid. must have a power flow attribute.}

\item{alpha}{The system tolerance.}

\item{PowerFlow}{The name of the graph attribute which contains the power flow over each line.}

\item{Link.Limit}{The attribute name that will be used when setting the line limits.}
}
\description{
This function sets the line limits to the tolerance value \eqn{\alpha} chosen by the user.
}
