% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlackRef.R
\name{SlackRefFunc}
\alias{SlackRefFunc}
\title{SlackRef}
\usage{
SlackRefFunc(g, name = "name", Bus.Order = "Bus.Order")
}
\arguments{
\item{g}{The igraph object representing the power grid}

\item{name}{The vertex attribute containing node names}

\item{Bus.Order}{The vertex attribute that contains the rank order for slack reference}
}
\description{
Finds the Slack node in each component of the power grid. A graph is put in and a data frame giving 
the slack reference for each component is returned
}
